remove necessity to save parameters
When we want to repeat the experiment with different values for some
parameters (e.g. cpar), we should keep the same values for the rest of
the parameters (bn, dn etc.). Before, we had to save the parameters and
load them in every loop. With this code we don't have to save them since
they are set before the loop that changes the value, and we manually set
the value of the changed parameter on params structure.